moving box api error into client module
