+++ examples/auth_developer.rs
@@ -3,9 +3,9 @@
 
 use rusty_box::{
     auth::auth_developer::DevAuth,
-    client::client::BoxClient,
+    client::{client::BoxClient, client_error::BoxAPIError},
     config::Config,
-    rest_api::{errors::error_api::BoxAPIError, users::users_api},
+    rest_api::users::users_api,
 };
 use std::env;
