+++ src/client.rs
@@ -1,2 +1,4 @@
 pub mod client;
+pub mod client_error;
+pub mod client_error_model;
 pub mod http_client;
