+++ src/client/client_error_model.rs
@@ -8,6 +8,20 @@
  * Generated by: https://openapi-generator.tech
  */
 
+#[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
+pub struct ClientErrorContextInfo {
+    /// More details on the error.
+    #[serde(rename = "message", skip_serializing_if = "Option::is_none")]
+    pub message: Option<String>,
+}
+
+impl ClientErrorContextInfo {
+    /// A free-form object that contains additional context about the error. The possible fields are defined on a per-endpoint basis. `message` is only one example.
+    pub fn new() -> ClientErrorContextInfo {
+        ClientErrorContextInfo { message: None }
+    }
+}
+
 /// ClientError : A generic error
 
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
@@ -25,7 +39,7 @@ pub struct BoxAPIErrorResponse {
     #[serde(rename = "message", skip_serializing_if = "Option::is_none")]
     pub message: Option<String>,
     #[serde(rename = "context_info", skip_serializing_if = "Option::is_none")]
-    pub context_info: Option<Box<super::client_error_context_info::ClientErrorContextInfo>>,
+    pub context_info: Option<Box<ClientErrorContextInfo>>,
     /// A URL that links to more information about why this error occurred.
     #[serde(rename = "help_url", skip_serializing_if = "Option::is_none")]
     pub help_url: Option<String>,
