--- src/http_client.rs
@@ -1,8 +1,0 @@
-//! HTTP client module.
-//! This module contains the `BaseHttpClient` trait, which is implemented by the `reqwest` module.
-// pub mod common;
-// pub mod reqwest;
-
-// pub use self::reqwest::ReqwestClient as HttpClient;
-
-// pub use common::{BaseHttpClient, Form, Headers, Query};
