+++ src/rest_api.rs
@@ -1,5 +1,4 @@
 //! Box REST API
 pub mod collaborations;
-pub mod errors;
 pub mod files;
 pub mod users;
