--- src/rest_api/errors/mod.rs
@@ -1,2 +1,0 @@
-pub mod error_api;
-pub mod model;
