--- src/rest_api/errors/model/client_error_context_info.rs
@@ -1,31 +1,0 @@
-/*
- * Box Platform API
- *
- * [Box Platform](https://box.dev) provides functionality to provide access to content stored within [Box](https://box.com). It provides endpoints for basic manipulation of files and folders, management of users within an enterprise, as well as more complex topics such as legal holds and retention policies.
- *
- * The version of the OpenAPI document: 2.0.0
- * Contact: devrel@box.com
- * Generated by: https://openapi-generator.tech
- */
-
-/// ClientErrorContextInfo : A free-form object that contains additional context about the error. The possible fields are defined on a per-endpoint basis. `message` is only one example.
-
-
-
-#[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
-pub struct ClientErrorContextInfo {
-    /// More details on the error.
-    #[serde(rename = "message", skip_serializing_if = "Option::is_none")]
-    pub message: Option<String>,
-}
-
-impl ClientErrorContextInfo {
-    /// A free-form object that contains additional context about the error. The possible fields are defined on a per-endpoint basis. `message` is only one example.
-    pub fn new() -> ClientErrorContextInfo {
-        ClientErrorContextInfo {
-            message: None,
-        }
-    }
-}
-
-
