--- src/rest_api/errors/model/mod.rs
@@ -1,2 +1,0 @@
-pub mod client_error;
-pub mod client_error_context_info;
